Move `isLocationEvent()` and `canForward()` to `EventUtils`
